+++ src/configs/jsx-a11y.ts
@@ -63,9 +63,9 @@ const defaultHoverOutHandlersRequiringOnBlur: Record<`on${string}`, true> = {
 
 export interface JsxA11yEslintConfigOptions extends UnConfigOptions<'jsx-a11y'> {
   /**
-   * [`eslint-plugin-jsx-a11y`](https://npmjs.com/eslint-plugin-jsx-a11y) plugin
+   * [`eslint-plugin-jsx-a11y-x`](https://npmjs.com/eslint-plugin-jsx-a11y-x) plugin
    * [shared settings](https://eslint.org/docs/latest/use/configure/configuration-files#configuring-shared-settings)
-   * that will be assigned to `jsx-a11y` property and applied to the specified `files` and `ignores`.
+   * that will be assigned to `jsx-a11y-x` property and applied to the specified `files` and `ignores`.
    */
   settings?: {
     attributes?: {
@@ -349,7 +349,7 @@ export const jsxA11yUnConfig: UnConfigFn<
             ...(isForLit
               ? pluginSettings
               : {
-                  'jsx-a11y': pluginSettings,
+                  'jsx-a11y-x': pluginSettings,
                 }),
           },
         }),